[UI] Ms2time cleanup, gtk version
git-svn-id: svn://svn.berlios.de/avidemux/branches/avidemux_2.6_branch_mean@5436 30592fff-cfd4-0310-bf0e-8235a3c7be26

+++ avidemux/gtk/ADM_userInterfaces/ADM_dialog/DIA_properties.cpp
@@ -19,7 +19,7 @@ void DIA_properties( void )
 {
 
  char text[80];
- uint16_t hh, mm, ss, ms;
+ uint32_t hh, mm, ss, ms;
  GtkWidget *dialog;
  uint8_t gmc, qpel,vop;
  uint32_t info=0;
